neater examples and fixed docs
