+++ examples/pagination_manual.rs
@@ -1,3 +1,6 @@
+//! This example shows how manual pagination works. It's what the raw API
+//! returns, but harder to use than an iterator or stream.
+
 use rspotify::client::SpotifyBuilder;
 use rspotify::oauth2::{CredentialsBuilder, OAuthBuilder};
 use rspotify::scopes;
@@ -31,7 +34,7 @@ async fn main() {
     //     .build()
     //     .unwrap();
     let oauth = OAuthBuilder::from_env()
-        .scope(scopes!("user-read-recently-played"))
+        .scope(scopes!("user-library-read"))
         .build()
         .unwrap();
 
@@ -44,8 +47,23 @@ async fn main() {
     // Obtaining the access token
     spotify.prompt_for_user_token().await.unwrap();
 
-    // Running the requests
-    let history = spotify.current_user_recently_played(10).await;
+    // Manual pagination. You may choose the number of items returned per
+    // iteration.
+    let limit = 50;
+    let mut offset = 0;
+    println!("Items:");
+    loop {
+        let page = spotify.current_user_saved_tracks(limit, offset).await.unwrap();
+        for item in page.items {
+            println!("* {}", item.track.name);
+        }
 
-    println!("Response: {:?}", history);
+        // The iteration ends when the `next` field is `None`. Otherwise, the
+        // Spotify API will keep returning empty lists from then on.
+        if page.next.is_none() {
+            break;
+        }
+
+        offset += limit;
+    }
 }
