+++ examples/with_refresh_token.rs
@@ -1,19 +1,19 @@
 //! Refresh tokens aren't meant to expire, so they can be used as a persistent
-//! authentication method without the need for user's interaction for
-//! oauth requests. You still need to authenticate the usual way at least
-//! once to obtain the refresh token, and you may need to obtain a new one
-//! if you change the required scope.
+//! authentication method without the need for user's interaction for oauth
+//! requests. You still need to authenticate the usual way at least once to
+//! obtain the refresh token, and you may need to obtain a new one if you change
+//! the required scope.
 //!
-//! The cache generated by `get_token` uses the refresh token under the hood
-//! to automatically authenticate the user. This example shows how it's done
-//! because sometimes it's not possible to use this cache file (a web server
-//! for example).
+//! The cache generated by `get_token` uses the refresh token under the hood to
+//! automatically authenticate the user. This example shows how it's done
+//! because sometimes it's not possible to use this cache file (a web server for
+//! example).
 //!
 //! *Note*: refresh tokens can actually expire, [as the OAuth2 spec
-//! indicates](https://tools.ietf.org/html/rfc6749#section-6),
-//! but this [hasn't actually happened in months with some
-//! tokens](https://github.com/felix-hilden/tekore/issues/86),
-//! so in the case of Spotify it doesn't seem to revoke them at all.
+//! indicates](https://tools.ietf.org/html/rfc6749#section-6), but this [hasn't
+//! actually happened in months with some
+//! tokens](https://github.com/felix-hilden/tekore/issues/86), so in the case of
+//! Spotify it doesn't seem to revoke them at all.
 
 use rspotify::client::{Spotify, SpotifyBuilder};
 use rspotify::model::Id;
